+++ ReduxCore/framework.php
@@ -69,7 +69,7 @@
             // Please update the build number with each push, no matter how small.
             // This will make for easier support when we ask users what version they are using.
 
-            public static $_version = '3.6.8';
+            public static $_version = '3.6.9';
             public static $_dir;
             public static $_url;
             public static $_upload_dir;