+++ ReduxCore/inc/fields/color_gradient/field_color_gradient.js
@@ -41,7 +41,7 @@
                             el.find( '#' + e.target.getAttribute( 'data-id' ) + '-transparency' ).removeAttr( 'checked' );
                         },
                         clear: function( e, ui ) {
-                            $( this ).val( ui.color.toString() );
+                            $( this ).val( '' );
                             redux_change( $( this ).parent().find( '.redux-color-init' ) );
                         }
                     }